requirement.txt
pygame==2.6.1
pypng==0.20220715.0
pyserial==3.5
RPi.GPIO==0.7.1
smbus==1.1.post2
smbus2==0.5.0
